+++ batch-stickers-gen.js
