+++ batch-stickers-templ.js
